requirements.txt
cryptocode==0.1
discord.py==2.4.0
pyppeteer==2.0.0
Pillow==11.0.0
audioop-lts==0.2.1